@@ -10,7 +10,7 @@
         }
     ],
     "require": {
-        "php": "^7.3|^8.0",
+        "php": "^7.4|^8.0",
         "laravel/framework": ">=8.0",
         "ext-json": "*"
     },
